feat: update zipperProductionStatusReportV2 to cast pricing fields to float for accurate calculations

+++ src/db/report/query/query.js
@@ -445,9 +445,9 @@ export async function zipperProductionStatusReportV2(req, res, next) {
                     CASE WHEN vodf.light_preference_name IS NOT NULL THEN ' ,' ELSE '' END,
                     COALESCE(vodf.light_preference_name, '')
                 ) AS other_details,
-                 vodf.md_price,
-                 vodf.mkt_company_price,
-                 vodf.mkt_party_price,
+                 vodf.md_price::float8,
+                 vodf.mkt_company_price::float8,
+                 vodf.mkt_party_price::float8,
                  vodf.is_price_confirmed
             FROM
                 zipper.v_order_details_full vodf
